remove ciruclar import, make AseTrajAdaptor static methods on Trajectory; add tests
